Prevent failing of PHPunit if document is null.
git-svn-id: https://svn.zib.de/opus4dev/framework/trunk@4461 0a49f362-4213-0410-86e3-e96cb6c4fbae

+++ tests/Opus/Search/DocumentAdapterTest.php
@@ -129,7 +129,9 @@ class Opus_Search_DocumentAdapterTest extends PHPUnit_Framework_TestCase {
      * @return void
      */
     public function tearDown() {
-        $this->_document->delete();
+        if (! is_null($this->_document) ) {
+            $this->_document->delete();
+        }
     }
 
     /**
